move socketSender to IOSession
